指摘していただいた箇所を修正
